Change to send email to SMTP server
Before, the program logged into a user smtp server, which required
authenticating as a user. This is because the server performed actions on
behalf of the authenticated user, which was not the goal for this program. This
changes it to send the email to an external SMTP server, which does not requrie
authentication.