Fix SMTP client to interface with external server (#32)
* Fix authentication issue

By connecting to an external outlook SMTP server, the program shouldn't need to
authenticate.

* Change to send email to SMTP server

Before, the program logged into a user smtp server, which required
authenticating as a user. This is because the server performed actions on
behalf of the authenticated user, which was not the goal for this program. This
changes it to send the email to an external SMTP server, which does not requrie
authentication.

* Remove unused libraries

Removed libraries are leftover from the mistake of authenticating with the
wrong SMTP server.

Co-authored-by: Reid Reininger <reidreininger@outlook.com>